requirements.txt
Django==5.0.1
django-extensions==3.2.3
django-ninja==1.3.0
python-decouple==3.8
ruff==0.1.9
